Fixed navigation updating in metadata
