+++ src/main/webapp/resources/js/pages/projects/settings/components/metadata/MetadataLayout.jsx
@@ -1,16 +1,6 @@
-import { Link, Router } from "@reach/router";
+import { Link } from "@reach/router";
 import { Menu, Space, Typography } from "antd";
 import React from "react";
-import { useSelector } from "react-redux";
-
-const MetadataFields = React.lazy(() => import("./MetadataFields"));
-const MetadataTemplateManager = React.lazy(() =>
-  import("./MetadataTemplateManager")
-);
-const MetadataTemplateMember = React.lazy(() =>
-  import("./MetadataTemplateMember")
-);
-const MetadataTemplates = React.lazy(() => import("./MetadataTemplates"));
 
 /**
  * Component for rendering the metadata fields and templates
@@ -18,37 +8,23 @@ const MetadataTemplates = React.lazy(() => import("./MetadataTemplates"));
  * @returns {JSX.Element}
  * @constructor
  */
-export default function MetadataLayout() {
-  const { canManage } = useSelector((state) => state.project);
-  const [selected, setSelected] = React.useState("fields");
+export default function MetadataLayout({ children, ...props }) {
   return (
-    <Space style={{ width: `100%` }} direction="vertical">
+    <>
       <Typography.Title level={2}>
         {i18n("project.settings.page.metadata")}
       </Typography.Title>
-      <Menu
-        mode="horizontal"
-        selectedKeys={[selected]}
-        onClick={(e) => setSelected(e.key)}
-      >
-        <Menu.Item key="fields" className="t-m-field-link">
-          <Link to="../metadata/fields">{i18n("MetadataFields.title")}</Link>
-        </Menu.Item>
-        <Menu.Item key="templates" className="t-m-template-link">
-          <Link to="../metadata/templates">
-            {i18n("ProjectMetadataTemplates.title")}
-          </Link>
-        </Menu.Item>
-      </Menu>
-      <Router>
-        <MetadataFields path="/fields" />
-        <MetadataTemplates path="/templates" />
-        {canManage ? (
-          <MetadataTemplateManager path="/templates/:id" />
-        ) : (
-          <MetadataTemplateMember path="/templates/:id" />
-        )}
-      </Router>
-    </Space>
+      <Space style={{ width: `100%` }} direction="vertical" size="large">
+        <Menu mode="horizontal" selectedKeys={[props["*"]]}>
+          <Menu.Item key="fields" className="t-m-field-link">
+            <Link to="fields">{i18n("MetadataFields.title")}</Link>
+          </Menu.Item>
+          <Menu.Item key="templates" className="t-m-template-link">
+            <Link to="templates">{i18n("ProjectMetadataTemplates.title")}</Link>
+          </Menu.Item>
+        </Menu>
+        <div>{children}</div>
+      </Space>
+    </>
   );
 }
